refactore utils module
