+++ .gitignore
@@ -7,3 +7,5 @@ ml_project.egg-info
 tests/file.sav
 file.sav
 *2022-05*/
+build/
+dist/
